Fully extract a NavBar component
Fully extract a NavBar component to handle navigation.

There are still several things that could be improved, but this vastly simplifies the main App component.
