+++ src/components/index.js
@@ -1,1 +1,2 @@
 export { default as ProjectCarousel } from './carousel/ProjectPreviewCarousel.vue'
+export { default as NavBar } from './nav/NavBar.vue'
